+++ crates/cgp-core/src/prelude.rs
@@ -3,4 +3,6 @@ pub use cgp_component::{
     define_components, delegate_components, derive_component, DelegateComponent, HasComponents,
 };
 pub use cgp_error::{CanRaiseError, HasErrorType};
-pub use cgp_field::{symbol, Char, HasField, HasFieldMut};
+pub use cgp_field::{
+    product, symbol, Char, Cons, Either, HasField, HasFieldMut, Nil, Product, Sum, Void,
+};