+++ crates/cgp-field-macro-lib/src/lib.rs
@@ -6,10 +6,12 @@
 */
 
 pub mod field;
+pub mod product;
 pub mod symbol;
 
 #[cfg(test)]
 mod tests;
 
 pub use field::derive_fields;
+pub use product::{make_product_expr, make_product_type, make_sum_type};
 pub use symbol::make_symbol;